+++ src/main/java/com/github/apoioSolidario/service/EventService.java
@@ -4,6 +4,7 @@ import com.github.apoioSolidario.dto.mapper.EventMapper;
 import com.github.apoioSolidario.dto.request.EventRequest;
 import com.github.apoioSolidario.dto.request.UpdateStatusRequest;
 import com.github.apoioSolidario.dto.response.EventResponse;
+import com.github.apoioSolidario.exception.AccessDeniedException;
 import com.github.apoioSolidario.model.Event;
 import com.github.apoioSolidario.model.Location;
 import com.github.apoioSolidario.model.Ong;
@@ -27,12 +28,14 @@ public class EventService {
     private final LocationRepository locationRepository;
     private final OngRepository ongRepository;
     private final EventMapper mapper;
+    private final TokenService tokenService;
 
-    public EventService(EventRepository repository, LocationRepository locationRepository, OngRepository ongRepository, EventMapper mapper) {
+    public EventService(EventRepository repository, LocationRepository locationRepository, OngRepository ongRepository, EventMapper mapper, TokenService tokenService) {
         this.repository = repository;
         this.locationRepository = locationRepository;
         this.ongRepository = ongRepository;
         this.mapper = mapper;
+        this.tokenService = tokenService;
     }
 
     public Page<EventResponse> findAll(Pageable pageable) {
@@ -41,16 +44,25 @@ public class EventService {
 
     public EventResponse findById(UUID id) {
         var entity = repository.findById(id).orElseThrow(() -> new EntityNotFoundException(id, "Event"));
+        if(!checkAccessEvent(entity.getOng().getOngId())){
+            throw new AccessDeniedException("Acesso negado ao recurso solicitado");
+        }
         return mapper.toObject(entity, EventResponse.class);
     }
 
     @Transactional()
     public EventResponse findByHandler(String handler) {
         var entity = repository.findByHandler(handler).orElseThrow(() -> new EntityNotFoundException(String.format("Campaingn com handler %s não encontrada", handler)));
+        if(!checkAccessEvent(entity.getOng().getOngId())){
+            throw new AccessDeniedException("Acesso negado ao recurso solicitado");
+        }
         return mapper.toObject(entity, EventResponse.class);
     }
 
     public EventResponse save(EventRequest request) {
+        if(!checkAccessEvent(request.getOngId())){
+            throw new AccessDeniedException("Acesso negado ao recurso solicitado");
+        }
         Ong ong = ongRepository.findById(request.getOngId()).orElseThrow(() -> new EntityNotFoundException(request.getOngId(), "Ong"));
         Location location = locationRepository.findById(request.getLocationId()).orElseThrow(() -> new EntityNotFoundException(request.getOngId(), "Location"));
         Event entity = mapper.toObject(request, Event.class);
@@ -65,6 +77,9 @@ public class EventService {
         Event entity = repository.findById(id).orElseThrow(
                 () -> new EntityNotFoundException(id, "event")
         );
+        if(!checkAccessEvent(entity.getOng().getOngId())){
+            throw new AccessDeniedException("Acesso negado ao recurso solicitado");
+        }
         Ong ong = ongRepository.findById(request.getOngId()).orElseThrow(() -> new EntityNotFoundException(request.getOngId(), "Ong"));
         Location location = locationRepository.findById(request.getLocationId()).orElseThrow(() -> new EntityNotFoundException(request.getOngId(), "Location"));
         entity.setTitle(request.getTitle());
@@ -82,6 +97,9 @@ public class EventService {
     public void deleteById(@Valid UUID id) {
         var entity = repository.findById(id).orElseThrow(() -> new EntityNotFoundException(id, "event")
         );
+        if(!checkAccessEvent(entity.getOng().getOngId())){
+            throw new AccessDeniedException("Acesso negado ao recurso solicitado");
+        }
         repository.delete(entity);
     }
 
@@ -89,6 +107,9 @@ public class EventService {
         Event entity = repository.findById(id).orElseThrow(
                 () -> new EntityNotFoundException(id, "event")
         );
+        if(!checkAccessEvent(entity.getOng().getOngId())){
+            throw new AccessDeniedException("Acesso negado ao recurso solicitado");
+        }
         entity.setStatus(request.getStatus());
         entity.setUpdatedAt(LocalDateTime.now());
         return mapper.toObject(repository.save(entity), EventResponse.class);
@@ -96,11 +117,20 @@ public class EventService {
 
     @Transactional
     public Page<EventResponse> finByOngId(UUID id, Pageable pageable) {
+        if(!checkAccessEvent(id)){
+            throw new AccessDeniedException("Acesso negado ao recurso solicitado");
+        }
         return mapper.toPage(repository.findByOng_OngId(id, pageable), EventResponse.class);
     }
 
     @Transactional
     public Page<EventResponse> finByLocationId(UUID id, Pageable pageable) {
         return mapper.toPage(repository.findByLocation_LocationId(id, pageable), EventResponse.class);
+    }
+
+    public boolean checkAccessEvent(UUID ongId) {
+        var logado = tokenService.getPrincipal();
+        Ong ongLogada = ongRepository.findByUser_UserId(logado.getUserId()).orElseThrow(()->new EntityNotFoundException(logado.getUserId(),"Ong"));
+        return tokenService.isAdmin() || ongLogada.getOngId().equals(ongId);
     }
 }
